+++ src/main/java/de/codescape/jira/plugins/scrumpoker/rest/mapper/SessionReferenceMapper.java
@@ -1,4 +1,4 @@
-package de.codescape.jira.plugins.scrumpoker.service;
+package de.codescape.jira.plugins.scrumpoker.rest.mapper;
 
 import com.atlassian.jira.issue.IssueManager;
 import com.atlassian.jira.issue.MutableIssue;
@@ -13,22 +13,22 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
- * Service that allows the transform a list of {@link ScrumPokerSession} elements into a {@link ReferenceListEntity}
- * that contains a {@link ReferenceEntity} element for every {@link ScrumPokerSession}. This service creates a model
- * that can be used and transferred as a REST resource and is optimized for a logic less templating mechanism.
+ * Service that allows the map a list of {@link ScrumPokerSession} elements to a {@link ReferenceListEntity} that
+ * contains a {@link ReferenceEntity} element for every {@link ScrumPokerSession}. This service creates a model that can
+ * be used and transferred as a REST resource and is optimized for a logic less templating mechanism.
  */
 @Component
-public class SessionReferenceTransformer {
+public class SessionReferenceMapper {
 
     private final IssueManager issueManager;
 
     @Autowired
-    public SessionReferenceTransformer(IssueManager issueManager) {
+    public SessionReferenceMapper(IssueManager issueManager) {
         this.issueManager = issueManager;
     }
 
     /**
-     * Transform a given list of {@link ScrumPokerSession} and a given estimation into a {@link ReferenceListEntity}.
+     * Map a given list of {@link ScrumPokerSession} and a given estimation to a {@link ReferenceListEntity}.
      *
      * @param scrumPokerSessions list of {@link ScrumPokerSession}
      * @param estimation         estimation